Extract risk rolling response module
